+++ Assets/Slua/LuaObject/Unity/Lua_UnityEngine_Random.cs
@@ -81,20 +81,20 @@ public class Lua_UnityEngine_Random : LuaObject {
 			#endif
 			#endif
 			int argc = LuaDLL.lua_gettop(l);
-			if(matchType(l,argc,1,typeof(int),typeof(int))){
-				System.Int32 a1;
+			if(matchType(l,argc,1,typeof(float),typeof(float))){
+				System.Single a1;
 				checkType(l,1,out a1);
-				System.Int32 a2;
+				System.Single a2;
 				checkType(l,2,out a2);
 				var ret=UnityEngine.Random.Range(a1,a2);
 				pushValue(l,true);
 				pushValue(l,ret);
 				return 2;
 			}
-			else if(matchType(l,argc,1,typeof(float),typeof(float))){
-				System.Single a1;
+			else if(matchType(l,argc,1,typeof(int),typeof(int))){
+				System.Int32 a1;
 				checkType(l,1,out a1);
-				System.Single a2;
+				System.Int32 a2;
 				checkType(l,2,out a2);
 				var ret=UnityEngine.Random.Range(a1,a2);
 				pushValue(l,true);
